Try to fix weird git issue
On macOS, git won't discard changes to a file, c_Simulation.f90,
despite calling

    git checkout -- .

If I specify the file with

    git checkout -- src/prmslib/misc/c_Simulation.f90

then the changes are discarded, but another file, c_simulation.f90
(note the lowercase "s"), is marked as changed. My hunch is that this
is a case-sensitive filesystem issue because of the similarity of
the two filenames. Perhaps a change was commited on Windows or a Mac
without a case-sensitive filesystem.

Not an issue on Linux.